+++ slurmweb/apps/agent.py
@@ -40,6 +40,7 @@ class SlurmwebAppAgent(SlurmwebWebApp, RFLTokenizedRBACWebApp):
         SlurmwebAppRoute(f"/v{get_version()}/ping", views.ping),
         SlurmwebAppRoute(f"/v{get_version()}/stats", views.stats),
         SlurmwebAppRoute(f"/v{get_version()}/jobs", views.jobs),
+        SlurmwebAppRoute(f"/v{get_version()}/jobs/past", views.jobs_past),
         SlurmwebAppRoute(f"/v{get_version()}/job/<int:job>", views.job),
         SlurmwebAppRoute(f"/v{get_version()}/nodes", views.nodes),
         SlurmwebAppRoute(f"/v{get_version()}/node/<name>", views.node),